+++ Windows/AndroidMic/packages.config
@@ -11,4 +11,6 @@
   <package id="NAudio.WinMM" version="2.1.0" targetFramework="net472" />
   <package id="System.Security.AccessControl" version="6.0.0" targetFramework="net472" />
   <package id="System.Security.Principal.Windows" version="5.0.0" targetFramework="net472" />
+
+	<package id="SharpAdbClient" version="2.3.23" targetFramework="net472" />
 </packages>
